remove old initializer
